refactor(patterns): share the note-button click helpers between the two default-app integration tests
The default-app integration tests find and click a button by its text through a
small family of helpers: a marker attribute, an in-page predicate that stamps the
resolved control (markNoteButton), a routine that clears those marks, the loop
that waits for a match and dispatches one trusted click
(settleAndClickNoteButton), and the three by-text and by-title entry points the
tests actually call. default-app.test.ts and reload/default-app-notebook.test.ts
each carried a byte-identical copy of that whole family, plus a findButtonWithText
finder that differed only in comment and brace style.

Move the family into a new note-button-helpers.ts beside the existing
cfc-browser-helpers.ts. Only the four call-site helpers are exported
(clickButtonWithText, clickButtonWithExactText, clickButtonWithTitle,
findButtonWithText); the marker attribute, the predicate, the clear routine, and
the click loop stay module-private. Both test files now import from it instead of
defining their own copies.

markNoteButton is serialized into the page by waitForCondition through its source
text, so it must reference nothing from module scope. Its comment now states that
constraint, matching the note the sibling cfc-browser-helpers.ts already carries
above its own serialized predicates.
